refactor(router): 重构路由配置为模块化结构
将路由配置按功能模块拆分为多个文件，提高可维护性
添加语义化颜色别名，优化仪表盘视觉指示器颜色
修复基础练习页HP奖励类型转换问题
添加简单版主程序入口文件

+++ lib/features/exercise/ui/base_exercise_page.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import '../../../core/theme/app_colors.dart';
+
 import '../../../core/constants/app_strings.dart';
-import '../../../core/theme/app_typography.dart';
 import '../../../core/router/app_router.dart';
-import '../../../shared/widgets/grid_background.dart';
-import '../../../shared/widgets/scanline_overlay.dart';
+import '../../../core/theme/app_colors.dart';
+import '../../../core/theme/app_typography.dart';
 import '../../../shared/providers/user_state_provider.dart';
 import '../../../shared/services/audio_service.dart';
+import '../../../shared/widgets/grid_background.dart';
+import '../../../shared/widgets/scanline_overlay.dart';
 import '../widgets/timer_display.dart';
 
 /// 动作执行基类
@@ -84,7 +85,7 @@ class _BaseExercisePageState extends ConsumerState<BaseExercisePage> with Ticker
   void _onComplete() {
     // 更新全局状态：奖励 HP 和金币
     final notifier = ref.read(userStateProvider.notifier);
-    notifier.addHp(widget.hpReward);
+    notifier.addHp(widget.hpReward.toDouble());
     notifier.addCoins(widget.coinReward);
     
     // 播放完成音效
